Rewrite how tests are executed and reported
Now I use a generic function to run the tests N times and data classes to store the results. This leads to less duplication.
I removed the min and max times, and report the average only.
Also, the output is now a markdown table for easy insertions in the Readme.